Comment out the username and password options in config/mongoid.yml
Why ? Because production settings don't work if they're nil, which is
the case on some platforms (mine for instance), leading to an application
that doesn't access properly to the database. The problem was a bit
tricky to identify since development settings work and the first error
returned is simply a nil value from a specific Setting.